+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 ## 1.6.5
+
+### Bug Fixes
+
+ * Fix Parameter '__frch_typeId_0' not found error